% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix.utils.R
\docType{package}
\name{Matrix.utils}
\alias{Matrix.utils}
\alias{Matrix.utils-package}
\title{Data.frame-like operations on sparse and dense \code{Matrix} objects,}
\description{
Implements cast, aggregate, and join for \code{\link{Matrix}} and matrix-like
objects.
}

