% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix.utils.R
\name{dMcast}
\alias{dMcast}
\title{Casts or pivots a long \code{data frame} into a wide sparse matrix.}
\usage{
dMcast(data, formula, fun.aggregate = "sum", value.var = NULL)
}
\arguments{
\item{data}{a data frame}

\item{formula}{casting \code{\link[stats]{formula}}, see details for
specifics.}

\item{fun.aggregate}{name of aggregation function.  Defaults to 'sum'}

\item{value.var}{name of column that stores values to be aggregated}
}
\description{
Similar in function to \code{\link[reshape2]{dcast}}, but produces a sparse 
\code{\link{Matrix}} as an output. Sparse matrices are beneficial for this 
application because such outputs are often very wide and sparse. Conceptually
similar to a \code{pivot} operation.
}
\details{
Casting formulas are slightly different than those in \code{dcast} and follow
the conventions of \code{\link{model.matrix}}. See \code{\link{formula}} for 
details.  Briefly, the left hand side of the \code{~} will be used as the 
grouping criteria.  This can either be a single variable, or a group of 
variables linked using \code{:}.  The right hand side specifies what the 
columns will be. Unlike \code{dcast}, using the \code{+} operator will append
the values for each variable as additional columns.  This is useful for 
things such as one-hot encoding.  Using \code{:} will combine the columns as 
interactions.
}
\seealso{
\code{\link[reshape]{cast}}

\code{\link[reshape2]{dcast}}

\code{\link[data.table]{dcast.data.table}}
}

