% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix.utils.R
\docType{package}
\name{Matrix.utils}
\alias{Matrix.utils}
\alias{Matrix.utils-package}
\title{Data.frame-Like Operations on Sparse and Dense \code{Matrix} Objects}
\description{
Implements data manipulation methods such as cast, aggregate, and merge/join
for \code{\link{Matrix}} and matrix-like objects.
}
